% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLColors.R
\docType{data}
\name{pl_palettes}
\alias{pl_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 33.
}
\usage{
pl_palettes
}
\description{
Use \code{names(pl_palettes)} to view list of palette names.
}
\keyword{datasets}
